refactored jsonapi files
