+++ composer.json
@@ -16,9 +16,9 @@
     },
     "autoload": {
 	"psr-4": {
-	    "ZK\\API\\": "api/",
 	    "ZK\\Controllers\\": "controllers/",
 	    "ZK\\Engine\\": [ "engine/", "engine/impl/" ],
+	    "ZK\\JsonApi\\": "jsonapi/",
 	    "ZK\\UI\\": "ui/"
 	},
 	"classmap": [
